# Conflicts:
#	src/software/firmware/srcs/pression.cpp

+++ src/software/firmware/includes/pression.h
@@ -30,4 +30,4 @@ int convertSensor2Pressure(uint16_t sensorValue);
  * used when in simulation mode)
  * @return         The current pressure in mmH20
  */
-int readPressureSensor(uint16_t centiSec);
+int16_t readPressureSensor(uint16_t centiSec);
